[ch] implement restart of app on changes -
Separate demo and test with env specific Docker file and add script to
allow for app to be restarted where there are any changes
(create,delete,modify) made whilst deloping the application.

This is more proof of concept as this will only work on a Linux OS with
`inotify-tools` installed, but as a proof of concept it _seems_ to work
gracefully.