Add file watcher (#20)
* [ch] enable sync between volumes -

In order to update the docker container with a file change, a docker
rebuild is required. This will stop the need to run a rebuild.

* [ch] set docker timeout to `0` -

The default timeout for docker `stop` and `restart` is `10` seconds.
Please see [Docker
Docs](https://docs.docker.com/compose/reference/restart/).

* [ch] remove cmd to make entry file executable -

This has been recently pointed out that this is not required as it is a
php file.

* [ch] add best practice to `ENTRYPOINT` -

It is considered more performant to pass `ENTRYPOINT` and `CMD` an array
of the required command.

* [f] revert `ENTRYPOINT` to `ServerStub.php`

* [ch] implement restart of app on changes -

Separate demo and test with env specific Docker file and add script to
allow for app to be restarted where there are any changes
(create,delete,modify) made whilst deloping the application.

This is more proof of concept as this will only work on a Linux OS with
`inotify-tools` installed, but as a proof of concept it _seems_ to work
gracefully.

* [ch] Add CLI flags -

Allow `h`, `--host`, `p`, `--port` to be passed to start the server.

* [a] using GoLang to watch local dev -

This will start up the ReactPHP server in a child process within the
GoLang script. Killing the script will kill the child process as
intended allowing for a better development environment.

* [ch] improve initial output code -

refs #17